Fix static analysis findings: mutex error handling, sprintf overflow, integer overflow
  - libdgr.c: Check pthread_mutex_lock return values in dgr_send and
    abort with diagnostic on failure; use oK() on unlock calls.
  - agents.c: Replace sprintf with snprintf to prevent buffer overrun
    on 128-byte filename buffer in agent_rotate_log.
  - metadata.c: Change parseCommandString capacity from int to size_t
    to prevent integer overflow in realloc size computation.

+++ bpv7/metadata/metadata.c
@@ -671,9 +671,9 @@ int isOnlyWhitespace(const char* str)
 /******************************************************************************/
 char **parseCommandString(const char *inputString, int *count)
 {
-	char  *str = myStrdup(inputString);
-	int    capacity = 5;
-	char **result = MTAKE(capacity * sizeof(char *));
+	char   *str = myStrdup(inputString);
+	size_t  capacity = 5;
+	char  **result = MTAKE(capacity * sizeof(char *));
 	*count = 0;
 	char *token = strtok(str, ",");
